+++ src/VtkWriter.cpp
@@ -8,7 +8,7 @@ namespace LBM {
 	void WriteVtk(const std::vector<double>& rho, 
 				   const std::vector<double>& ux, 
 				   const std::vector<double>& uy, 
-				   const int Nx, const int Ny, 
+				   const size_t Nx, const size_t Ny, 
 				   const std::string& file_path, 
 				   const std::string& pv_title) {
